add not dislike again

+++ src/main/java/com/group4/HaUISocialMedia_server/service/impl/LikeServiceImpl.java
@@ -107,6 +107,8 @@ public class LikeServiceImpl implements LikeService {
         if (post == null)
             return false;
 
+        if(likeRepository.findByUserAndPost(postId, user.getId()) == null)
+            return false;
         likeRepository.deleteByIdPost(postId, user.getId());
 
         Notification oldNotification = notificationRepository.getOldLikeNotification(post.getOwner().getId(), postId);
